requirements.txt
Django==2.2
numpy==1.20.3
pandas==1.3.4
scikit-learn==1.0.1
django-storages[azure]
pyodbc
django-mssql-backend
django-pyodbc-azure